module_exists moved to helper.py
